Move `expand_paths` into a job
The reason is that we want to expand environment variables before doing the validation of the final values;
This will also ensure that if the job configuration changes because the referenced environment variables
have changed, then Serge will properly disable optimizations on next run.

Due to this change, the job object is blessed before the hash calculation is done, so this will change
job hashes in all tests; updated test data will be committed separately.