feature_branch: properly handle orphaned strings in master branch
Prevously, when the string was translated and then deleted
from the master branch, feature_branch plugin would reuse the
such translations, but wouldn't expose the string in translation
interchange file of a feature branch if the same string would reappear
in such a feature branch.

With this change, the string that disappeared from master branch
and resurfaced in a feature branch will be editable in the context
of that branch.