Stabilize the translation selection with multiple variants
Previously, if there would be multiple translation
variants with the same fitness, the first one would be used;
however, due to the fact that access to keys/values of a hash
is random, this would mean a random translation would be used
which is not good for consistency, and would fail repeated tests.

With this change, translations sharing the best fitness are
accumulated in an array, and then the first one in alphabetical
sort order is pickedm which stabilizes the output.