Refactor feature_branch plugin logic
The previous implementation loaded its own cache from the database
and resolved duplicate translations in slave jobs in its own way,
not consistent with the regular Serge behavior.

The new implementation is much simpler; it uses the log_translation
callback to save translations for the given file/key/string/context
in the master job, and then uses these exact translations in
slave jobs, ensuring the translations for exact matches are identical.