`transform` plugin: guessed translations are now returned as fuzzy
Previously plugin would determine the fuzzy bit logic from the
job settings. However, due to the nature of the plugin, it's much
wiser to unconditionally raise the fuzzy flag to force the review
of such strings, because punctuation and letter casing may require
language-specific adjustments that this plugin can't figure out
by itself.