Engine: normalize string before running the segmentation plugin
If `normalize_strings` option is turned on, and if a segmentation
plugin is enabled at the same time, we wouldn't be normalizing
strings until after they pass through the segmentation plugin.

However, this kind of normalization is supposed to happen before
segmentation, as it can help with segmenting sentences
(and avoid unnecessary segmentation when long sentences are broken
into multiple lines).